refactor: cabinet dto, serializer to modular #56
